Separate types, improve formatting
