+++ src/flightradar_data/types/__init__.py
@@ -1,0 +1,9 @@
+from src.flightradar_data.types.airport import Airport
+from src.flightradar_data.types.flight import Flight
+from src.flightradar_data.types.runway import Runway
+
+__all__ = [
+    'Airport',
+    'Flight',
+    'Runway'
+]
