+++ src/flightradar_data/types/airport.py
@@ -1,90 +1,14 @@
 import logging
 import time
-import traceback
 from dataclasses import dataclass, field
 from typing import Any, Literal
 
 import requests
 
+from src.flightradar_data.types.flight import Flight
+from src.flightradar_data.types.runway import Runway
 
 _logger = logging.getLogger(__name__)
-
-
-@dataclass(unsafe_hash=True)
-class Flight:
-    aircraft_model: str
-    airline: str
-    destination_iata: str
-    destination_icao: str
-    flight_number: str
-    long_aircraft_name: str
-    origin_iata: str
-    origin_icao: str
-    scheduled_arrival: int
-    scheduled_departure: int
-
-    @classmethod
-    def new_flight(cls, airport_iata: str, airport_icao: str, /, *, flight_data: dict[str, Any],
-                   is_arrival: bool) -> 'Flight':
-        flight_info = flight_data['flight']
-        if is_arrival:
-            origin_info = flight_info['airport']['origin']
-            origin_iata = None if not origin_info else origin_info['code']['iata']
-            origin_icao = None if not origin_info else origin_info['code']['icao']
-            destination_iata = airport_iata
-            destination_icao = airport_icao
-        else:
-            destination_info = flight_info['airport']['destination']
-            destination_iata = None if not destination_info else destination_info['code']['iata']
-            destination_icao = None if not destination_info else destination_info['code']['icao']
-            origin_iata = airport_iata
-            origin_icao = airport_icao
-
-        airline = None if not flight_info['airline'] else flight_info['airline']['name']
-        aircraft_model = None if not flight_info['aircraft'] else flight_info['aircraft']['model']['code']
-        aircraft_long_name = None if not flight_info['aircraft'] else flight_info['aircraft']['model']['text']
-        try:
-            return Flight(
-                aircraft_model=aircraft_model,
-                airline=airline,
-                destination_iata=destination_iata,
-                destination_icao=destination_icao,
-                flight_number=flight_info['identification']['number']['default'],
-                long_aircraft_name=aircraft_long_name,
-                origin_iata=origin_iata,
-                origin_icao=origin_icao,
-                scheduled_arrival=flight_info['time']['scheduled']['arrival'],
-                scheduled_departure=flight_info['time']['scheduled']['departure']
-            )
-        except TypeError as e:
-            _logger.critical(traceback.format_exc())
-            _logger.error(flight_info)
-
-    def json(self) -> dict[str, str | int]:
-        return self.__dict__
-
-    def csv(self) -> tuple[list[str], list[str | int]]:
-        return list(self.__dict__.keys()), list(self.__dict__.values())
-
-
-@dataclass(unsafe_hash=True)
-class Runway:
-    name: str
-    length_in_m: int
-    length_in_ft: int
-    surface: str
-
-    @classmethod
-    def new_runway(cls, runway_info: dict[str, Any]) -> 'Runway':
-        return Runway(
-            name=runway_info['name'],
-            length_in_m=runway_info['length']['m'],
-            length_in_ft=runway_info['length']['ft'],
-            surface=runway_info['surface']['name']
-        )
-
-    def json(self) -> dict[str, str | int]:
-        return self.__dict__
 
 
 @dataclass
@@ -111,7 +35,7 @@ class Airport:
             _logger.info(f'Retrying request in 2 seconds (attempt {retry}/5)...')
             time.sleep(2)
             if retry < 5:
-                self.update_runways(airport_endpoint, headers, retry + 1)
+                return self.update_runways(airport_endpoint, headers, retry + 1)
             else:
                 resp.raise_for_status()
         if not resp.status_code == 200:
@@ -136,13 +60,13 @@ class Airport:
             page = 2
             while available_flights >= current_flights:
                 _logger.info(f'Fetching {type} for {self.name} ({self.iata}) from {endpoint!r}... '
-                             f'Currently on page {page}/{int(available_flights/page_size) + 1}')
+                             f'Currently on page {page}/{int(available_flights / page_size) + 1}')
                 resp_json = self._fetch_next_page(endpoint, type, headers, page=page)
                 self.flights |= self._parse_flights(resp_json['data'], type)
                 current_flights += page_size
                 page += 1
 
-    def _fetch_next_page(self, endpoint: str, type: Literal['arrivals', 'departures'], headers: dict[str, Any], page: int, retry: int = 1) -> dict[Any, Any]:
+    def _fetch_next_page(self, endpoint: str, type: Literal['arrivals', 'departures'], headers: dict[str, Any],  page: int, retry: int = 1) -> dict[Any, Any]:
         endpoint = f'{endpoint}&page={page}'
         if page % 7 == 0:
             time.sleep(1.5)
@@ -166,7 +90,8 @@ class Airport:
 
     def _parse_flights(self, flights_data: list[dict[str, Any]], type: Literal['arrivals', 'departures']) -> set[Flight]:
         is_arrival = type == 'arrivals'
-        return {Flight.new_flight(self.iata, self.icao, flight_data=flight, is_arrival=is_arrival) for flight in flights_data}
+        return {Flight.new_flight(self.iata, self.icao, flight_data=flight, is_arrival=is_arrival) for flight in
+                flights_data}
 
     def json(self) -> dict[str, str | int | dict[str, str | int]]:
         return {
