+++ src/flightradar_data/utils/__init__.py
@@ -1,0 +1,5 @@
+from src.flightradar_data.utils.logging import setup_logger
+
+__all__ = [
+    'setup_logger'
+]
